Reverse making config parser functions const
These functions can't *actually* be const because in checking that a
particular key exists, it's section is accessed using the [] operator,
which is not const. Now, obviously the operation is *actually* const
because we check for the existance of the section before accessing it,
but the compiler doesn't know that so the function can't be const and
therefore none of the config parser's functions can.