Radical changes to how punches are handled and minor refactoring
Radical changes:
State now keeps track of the reference for a single punch which is set
when a punch is inputted and the conditions for selecting a new punch
are met. Punch is now a full-fledged class.

Minor refactoring:
Removed unnecessary prefixes such as is or get from
getter functions. Java-isms like these make it very tedious and verbose
to get anything from objects and they're aren't usually necessary.