More insane changes to punching
This will hopefully be the last major design change...

The idea now is that, since this game is 2D and all, notions of lead or
rear hits only exists at the animation level, and the player nor the
game need concern themselves with it. Players now throw jabs with the
jab control, throw crosses by pressing the power control and throw
hooks and uppers (decided at the animation level since I consider they
serve the same purpose) by holding the power button.