Streamline punches
We don't need to know weather a punch is body blow, for instance,
because that can be deduced from which hurtbox it connects to---body
shots aim low and hit the body hurtbox.

Punches now also require a specific action to be thrown, which will
allow to make distinctions between long pressing and just pressing
(long pressing remains unimplemented).