Simplify input
Instead of having weird actions like long_press, it's easier for
punches to require inputs to be in between a certain window of held
time to be thrown.

Double presses remain unimplemented. Possibly because they don't need
to: maybe checking for a double press can be accomplished by checking
the discrepancy between the ttl of a buffer with a press and
DOUBLE_PRESS_T when a press is detected. Depending on how verbose this
turns out to be, a function in input to check this might be in order.