Make base resolution 640x480
A higher base resolution makes fonts look better without having to mess
around with the code.

The state of the game right now is that art is done at a resolution of
160x120, which makes it easier for unskilled people such as myself to
draw half-way decent sprites. This art is exported as assets scaled up
4 times, making sprites 128x128 and background textures 640x480. This
means that the minimum resolution is 160x120, but 640x480 makes scaling
fonts less of a headache.