Major cleanup operation
main.hpp was destroyed, dt is now passed around.

Debugging is now compartmentalized in many different methods which
allows the config file to not be passed around, rather main can decide
if the debugging methods should be executed. This helps readability big
time because debugging functions are now neatly tucked into the bottom
of the implementation file.