Dodges: Replace dodges with single "smart" dodge
The purpose of the various dodges overlapped. I though a better way to
handle dodging would be to have a single dodge that animated according
to circumstance: a close-range dodge results in a duck, a mid-range one
in a slip and a long-range one in a pull. Animation-wise at least.