Rewrite
Ok, story time.

I finally got tired with how ugly and coupled my classes were so I
decided it was time for a big rewrite and clean-up. However, while I was
working on it, I noticed how illogical the whole structure of the
project was so I also took the time to reimplement a lot of behaviour.
I also, along the way, came up with lots of input and controls changes,
along with any number of undocumented behaviour changes.

I'm sure this rewrite still needs some work but I think I got it to the
point the previous code was so I'm finally resuming commits.